+++ app/src/main/java/com/simonag/simonag/model/Pertanggal.java
@@ -1,5 +1,7 @@
 package com.simonag.simonag.model;
 
+import java.math.BigInteger;
+
 /**
  * Created by diditsepiyanto on 6/18/17.
  */
@@ -8,9 +10,9 @@ public class Pertanggal {
     int id;
     String tanggal;
     int total_aktivitas;
-    int total_biaya;
+    BigInteger total_biaya;
 
-    public Pertanggal(int id, String tanggal, int total_aktivitas, int total_biaya) {
+    public Pertanggal(int id, String tanggal, int total_aktivitas, BigInteger total_biaya) {
         this.id = id;
         this.tanggal = tanggal;
         this.total_aktivitas = total_aktivitas;
@@ -41,11 +43,11 @@ public class Pertanggal {
         this.total_aktivitas = total_aktivitas;
     }
 
-    public int getTotal_biaya() {
+    public BigInteger getTotal_biaya() {
         return total_biaya;
     }
 
-    public void setTotal_biaya(int total_biaya) {
+    public void setTotal_biaya(BigInteger total_biaya) {
         this.total_biaya = total_biaya;
     }
 }
